# escrow-bounty/doc/swagger/escrow_bounty.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "application.proto",
    "version": "version not set"
  },
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/bounties": {
      "get": {
        "summary": "获取悬赏大厅列表",
        "operationId": "EscrowBountyService_ListBounties",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbListBountiesResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "pageId",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          }
        ],
        "tags": [
          "EscrowBountyService"
        ]
      },
      "post": {
        "summary": "雇主发布悬赏",
        "operationId": "EscrowBountyService_CreateBounty",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbCreateBountyResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/pbCreateBountyRequest"
            }
          }
        ],
        "tags": [
          "EscrowBountyService"
        ]
      }
    },
...
    "/v1/bounties/{bountyId}": {
      "delete": {
        "summary": "雇主删除悬赏（仅 PENDING/FAILED 状态可删除）",
        "operationId": "EscrowBountyService_DeleteBounty",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbDeleteBountyResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "bountyId",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          }
        ],
        "tags": [
          "EscrowBountyService"
        ]
      }
    },
    "/v1/bounties/{bountyId}/accept": {
      "post": {
        "summary": "猎人申请接单",
        "operationId": "EscrowBountyService_AcceptBounty",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbAcceptBountyResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "bountyId",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/EscrowBountyServiceAcceptBountyBody"
            }
          }
        ],
        "tags": [
          "EscrowBountyService"
        ]
      }
    },
    "/v1/bounties/{bountyId}/approve": {
      "post": {
        "summary": "雇主审核通过",
        "operationId": "EscrowBountyService_ApproveBounty",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbApproveBountyResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "bountyId",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/EscrowBountyServiceApproveBountyBody"
            }
          }
        ],
        "tags": [
          "EscrowBountyService"
        ]
      }
    },
    "/v1/bounties/{bountyId}/cancel": {
      "post": {
        "summary": "雇主取消悬赏并退款",
        "operationId": "EscrowBountyService_CancelBounty",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbCancelBountyResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "bountyId",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/EscrowBountyServiceCancelBountyBody"
            }
          }
        ],
        "tags": [
          "EscrowBountyService"
        ]
      }
    },
    "/v1/bounties/{bountyId}/comments": {
      "get": {
        "operationId": "EscrowBountyService_ListComments",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbListCommentsResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "bountyId",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          }
        ],
        "tags": [
          "EscrowBountyService"
        ]
      },
      "post": {
        "operationId": "EscrowBountyService_CreateComment",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbCreateCommentResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "bountyId",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/EscrowBountyServiceCreateCommentBody"
            }
          }
        ],
        "tags": [
          "EscrowBountyService"
        ]
      }
    },
    "/v1/bounties/{bountyId}/complete": {
      "post": {
        "summary": "雇主确认完成并打款",
        "operationId": "EscrowBountyService_CompleteBounty",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbCompleteBountyResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "bountyId",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/EscrowBountyServiceCompleteBountyBody"
            }
          }
        ],
        "tags": [
          "EscrowBountyService"
        ]
      }
    },
    "/v1/bounties/{bountyId}/confirm": {
      "post": {
        "summary": "雇主选定猎人",
        "operationId": "EscrowBountyService_ConfirmHunter",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbConfirmHunterResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "bountyId",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/EscrowBountyServiceConfirmHunterBody"
            }
          }
        ],
        "tags": [
          "EscrowBountyService"
        ]
      }
    },
    "/v1/bounties/{bountyId}/invitations": {
      "post": {
        "summary": "悬赏者邀请猎人接单",
        "operationId": "EscrowBountyService_CreateInvitation",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbCreateInvitationResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "bountyId",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/EscrowBountyServiceCreateInvitationBody"
            }
          }
        ],
        "tags": [
          "EscrowBountyService"
        ]
      }
    },
    "/v1/bounties/{bountyId}/messages": {
      "get": {
        "operationId": "EscrowBountyService_ListMessages",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbListMessagesResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "bountyId",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          }
        ],
        "tags": [
          "EscrowBountyService"
        ]
      },
      "post": {
        "summary": "悬赏内聊天（原有）",
        "operationId": "EscrowBountyService_SendMessage",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbSendMessageResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "bountyId",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/EscrowBountyServiceSendMessageBody"
            }
          }
        ],
        "tags": [
          "EscrowBountyService"
        ]
      }
    },
    "/v1/bounties/{bountyId}/reject": {
      "post": {
        "summary": "雇主审核拒绝",
        "operationId": "EscrowBountyService_RejectBounty",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbRejectBountyResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "bountyId",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/EscrowBountyServiceRejectBountyBody"
            }
          }
        ],
        "tags": [
          "EscrowBountyService"
        ]
      }
    },
    "/v1/bounties/{bountyId}/submit": {
      "post": {
        "summary": "猎人提交工作成果",
        "operationId": "EscrowBountyService_SubmitBounty",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbSubmitBountyResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "bountyId",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/EscrowBountyServiceSubmitBountyBody"
            }
          }
        ],
        "tags": [
          "EscrowBountyService"
        ]
      }
    },
    "/v1/bounties/{id}": {
      "get": {
        "summary": "获取单个悬赏详情",
        "operationId": "EscrowBountyService_GetBounty",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbGetBountyResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          }
        ],
        "tags": [
          "EscrowBountyService"
        ]
      }
    },
    "/v1/comments": {
      "get": {
        "operationId": "EscrowBountyService_ListUserComments",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbListUserCommentsResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "username",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "pageId",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          }
        ],
        "tags": [
          "EscrowBountyService"
        ]
      }
    },
    "/v1/comments/{commentId}": {
      "delete": {
        "operationId": "EscrowBountyService_DeleteComment",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbDeleteCommentResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "commentId",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          }
        ],
        "tags": [
          "EscrowBountyService"
        ]
      }
    },
    "/v1/conversations": {
      "get": {
        "operationId": "EscrowBountyService_ListConversations",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbListConversationsResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "EscrowBountyService"
        ]
      },
      "post": {
        "operationId": "EscrowBountyService_GetOrCreateConversation",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbGetOrCreateConversationResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/pbGetOrCreateConversationRequest"
            }
          }
        ],
        "tags": [
          "EscrowBountyService"
        ]
      }
    },
    "/v1/conversations/unread": {
      "get": {
        "operationId": "EscrowBountyService_GetUnreadCounts",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbGetUnreadCountsResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "EscrowBountyService"
        ]
      }
    },
    "/v1/conversations/{conversationId}": {
      "delete": {
        "operationId": "EscrowBountyService_DeleteConversation",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbDeleteConversationResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "conversationId",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          }
        ],
        "tags": [
          "EscrowBountyService"
        ]
      }
    },
    "/v1/conversations/{conversationId}/messages": {
      "get": {
        "operationId": "EscrowBountyService_ListPrivateMessages",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbListPrivateMessagesResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "conversationId",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          },
          {
            "name": "beforeId",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "int64"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          }
        ],
        "tags": [
          "EscrowBountyService"
        ]
      }
    },
...
  },
  "definitions": {
    "EscrowBountyServiceAcceptBountyBody": {
      "type": "object",
      "properties": {
        "hunterAccountId": {
          "type": "string",
          "format": "int64"
        }
      },
      "title": "申请接单相关请求/响应"
    },
...
    "pbAcceptBountyResponse": {
      "type": "object",
      "properties": {
        "application": {
          "$ref": "#/definitions/pbBountyApplication"
        }
      }
    },
    "pbApproveBountyResponse": {
      "type": "object",
      "properties": {
        "bounty": {
          "$ref": "#/definitions/pbBounty"
        }
      }
    },
...
    "pbCreateBountyRequest": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "rewardAmount": {
          "type": "string",
          "format": "int64"
        },
        "employerAccountId": {
          "type": "string",
          "format": "int64"
        },
        "employerUsername": {
          "type": "string",
          "title": "可选；未提供则从 JWT token 中获取"
        },
        "deadlineTimestamp": {
          "type": "string",
          "format": "int64",
          "title": "截止时间（秒），可选"
        }
      },
      "title": "悬赏相关请求/响应"
    },
    "pbCreateBountyResponse": {
      "type": "object",
      "properties": {
        "bounty": {
          "$ref": "#/definitions/pbBounty"
        }
      }
    },
...
    "pbDeleteBountyResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        }
      }
    },
...
    "pbListBountiesResponse": {
      "type": "object",
      "properties": {
        "bounties": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/pbBounty"
          }
        }
      }
    },
...
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}